Added information about drops and monsters
Since guaranteed drop information is in the checklist for the Red-Eyed Royal Guard, I thought the guaranteed drop information for the Pus of Man also needed to be on the list. And I add that the player can get an item from Mimic without killing it if they use an Undead Hunter Charm on it.