Logan's Scroll
Evidence is clear: NPC is wearing court sorcerer gear. And the lore text on Logan's Scroll, the Court Sorcerer Set, and the Court Sorcerer Staff tell the story that the scroll belonged to the court sorcerers who claim the legacy of Big Hat Logan by imitating his spells and staff.