Rearrange Greirat questline
They way in which Greirat's questline is currently presented in the sheet is somewhat suboptimal:
- It tells you to free him from his cell before you can even have the key.
- It tells you that his gesture is gotten later, while it can actually be obtained immediately after presenting the bone.
- It tells you to send him away right after killing a boss, making it a long stretch before he returns with the next boss kill.
- It tells you about sending him to his death long before his ashes can be retrieved. If he is sent off at this point, his wares will be unavailable for a long time.

Some of this is improved by moving and rewording some entries. But I cannot, for example, simply remove an existing checkbox from the beginning of Road of Sacrifices and add one to the end of Undead Settlement instead, as that would mess up the counter for Road of Sacrifices. Therefore, I add some new entries at the points where I think this questline should be advanced, and link them to the existing checkboxes.