Add more information on the Altar of Sunlight
I tried looking for this entry but couldn't find it. I added the name of the altar rather than just referring to it as 'the altar'.

In addition to receiving the Praise the Sun gesture I also added the information of being able to hand in Sunlight Medals for the Warrior of Sunlight covenant. If this is too detailed let me know and I'll resubmit without that info.

I personally think this makes for a good addition because it allows people to find this altar more easily in the document.