Fix order of playthrough instructions for the Drag Heap
The description of `playthrough_21_33` says to drop down *here*, presumably referring to the place where you picked up the Homeward Bone in `playthrough_21_32`. The instructions for the Angel Summoner shouldn't be placed in between those two because it makes it sound like you are meant to drop down from the roof with the Angel Summoner, which would be certain death.